[aleph-render] Remove redundant cast

+++ lib/plugins/aleph-render/src/upload_manager/mod.rs
@@ -278,7 +278,7 @@ impl TextureUploadSource {
     ) -> BufferToTextureCopyRegion {
         BufferToTextureCopyRegion {
             src: ImageDataLayout {
-                offset: self.offset as u64,
+                offset: self.offset,
                 row_pitch: self.desc.aligned_width(),
                 extent: self.desc.extent(),
             },
